Refactor NEBM
